+++ __init__.py
@@ -15,15 +15,16 @@ debug = False
 
 
 from . utils import config
-config.new_config()
+# config.new_config()
 
 
-# FOLDERS WITH MODULES THAT CONTAIN <register> AND <unregister> FUNCTIONS (Operators, UI Panels, etc).
+# FOLDERS WITH MODULES THAT CONTAIN <register> OR <unregister> FUNCTIONS (Operators, UI Panels, etc).
 sub_folders = [
-    'ui',  # Contains the addon preferences class (register first, unregister last).
+    'config',   # Contains the addon preferences (register first and unregister last)
     'operators_internal',
     'operators',
-    'config',
+    'customize',
+    'ui',
 ]
 
 from . utils import mod_utils