+++ customize/resources.py
@@ -3,14 +3,13 @@ import os
 
 from .. utils import (
     addon,
-    debug,
     decorators,
     files,
     paths,
 )
 
 @decorators.abstract_class_attributes('SOURCE_PATH', 'TARGET_PATH')
-class StudioResource():
+class StudioResources():
     @classmethod
     def load(cls):
         files.copy_files(cls.SOURCE_PATH, cls.TARGET_PATH)
@@ -22,16 +21,31 @@ class StudioResource():
         bpy.context.preferences.studio_lights.refresh()
 
 
-class Matcaps(StudioResource):
+class Matcaps(StudioResources):
     SOURCE_PATH = paths.ResourcePaths.matcaps
     TARGET_PATH = paths.BlenderPaths.matcaps
 
 
-class HDRI(StudioResource):
+class HDRIs(StudioResources):
     SOURCE_PATH = paths.ResourcePaths.hdri
     TARGET_PATH = paths.BlenderPaths.hdri
 
 
-class StudioLights(StudioResource):
+class StudioLights(StudioResources):
     SOURCE_PATH = paths.ResourcePaths.studiolights
     TARGET_PATH = paths.BlenderPaths.studiolights
+
+
+def register():
+    prefs = addon.prefs()
+
+    Matcaps.load()      if prefs.matcaps       else Matcaps.unload()
+    HDRIs.load()        if prefs.hdris         else HDRIs.unload()
+    StudioLights.load() if prefs.studio_lights else StudioLights.unload()
+
+
+def unregister():
+    if isinstance(bpy.context.space_data, bpy.types.SpacePreferences):
+        Matcaps.unload()
+        HDRIs.unload()
+        StudioLights.unload()