+++ utils/debug.py
@@ -2,5 +2,4 @@ from . import addon
 
 def msg(*messages, space=0):
     if addon.debug():
-        message = [str(message).ljust(space, ' ') for message in messages]
-        print(*message)
+        print(*[str(element).ljust(space, ' ') for element in messages])