+++ utils/keymap_utils.py
@@ -12,25 +12,36 @@ class KeymapGroup(ABC):
     def __init__(self, name):
         self.name = name
         self.keymap_list = []
+        self.km = None
+        self.kmi = None
         self.error_count = 0
     
-    # type = key, value = event
-    def add(self, km, idname, type, value, ctrl=False, alt=False, shift=False):
-        kmi = km.keymap_items.new(idname, type, value, ctrl=ctrl, alt=alt, shift=shift)
-        self.keymap_list.append((km, kmi))
+    # type=key, value=event
+    def add(self, idname, type, value, ctrl=False, alt=False, shift=False):
+        self.kmi = self.km.keymap_items.new(idname, type, value, ctrl=ctrl, alt=alt, shift=shift)
+        self.keymap_list.append((self.km, self.kmi))
         debug.msg(f'    {idname}{" CTRL" if ctrl else ""}{" ALT" if alt else ""}{" SHIFT" if shift else ""} {type} {value}')
-        return kmi
     
+    def prop(self, attr, val):
+        try:
+            setattr(self.kmi.properties, attr, val)
+        except Exception as e:
+            self.error_count += 1
+            print(f'    ARMORED-Toolkit WARNING: {e}, check for typos?')
+
     @abstractmethod
     def register():
         pass
-        # raise NotImplementedError
     
     def unregister(self):
-        if not self.keymap_list:
-            debug.msg(self._formatted_name, '[Nothing to Unregister]', space=22)
+        if self._empty_keymap_list(): 
             return
 
+        self._unregister_keymaps()
+        self.keymap_list.clear()
+        self.disabled_message()
+
+    def _unregister_keymaps(self):
         self.error_count = 0
         for km, kmi in self.keymap_list:
             try:
@@ -39,13 +50,17 @@ class KeymapGroup(ABC):
             except Exception as e:
                 self.error_count += 1
                 debug.msg(e, 'Probably an F2 Addon Exception')
-        
-        self.keymap_list.clear()
-        self.error_count = 0
-        debug.msg(f'Disabled {self._formatted_name} {self._error_msg()}')
+
+    def _empty_keymap_list(self):
+        if not self.keymap_list:
+            debug.msg(self._formatted_name, '[Nothing to Unregister]', space=22)
+            return True
 
     def enabled_message(self):
         debug.msg(f'Enabled {self._formatted_name} {self._error_msg()}')
+    
+    def disabled_message(self):
+        debug.msg(f'Disabled {self._formatted_name} {self._error_msg()}')
 
     def _error_msg(self):
         return '' if not self.error_count else f'with {self.error_count} errors.'
@@ -53,10 +68,3 @@ class KeymapGroup(ABC):
     @property
     def _formatted_name(self):
         return self.name.replace("_", " ").title()
-
-    def kmi_props(self, kmi_props, attr, value):
-        try:
-            setattr(kmi_props, attr, value)
-        except Exception as e:
-            self.error_count += 1
-            print(f'    ARMORED-Toolkit WARNING: {e}, check for typos?')